dependencies/requirements.txt
python-dotenv==1.0.1
click==8.1.7
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
pypdf==5.0.1
rapidfuzz==3.9.7
chardet==5.2.0
rich==13.8.1
